+++ backend/store/etcd/token_store.go
@@ -9,6 +9,7 @@ import (
 	"github.com/coreos/etcd/clientv3"
 	jwt "github.com/dgrijalva/jwt-go"
 	"github.com/sensu/sensu-go/api/core/v2"
+	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/types"
 )
 
@@ -72,12 +73,13 @@ func (s *Store) RevokeTokens(claims ...*v2.Claims) error {
 
 // GetToken gets a Claims.
 func (s *Store) GetToken(subject, id string) (*types.Claims, error) {
-	resp, err := s.client.Get(context.TODO(), getTokenPath(subject, id), clientv3.WithLimit(1))
+	key := getTokenPath(subject, id)
+	resp, err := s.client.Get(context.TODO(), key, clientv3.WithLimit(1))
 	if err != nil {
 		return nil, err
 	}
 	if len(resp.Kvs) != 1 {
-		return nil, fmt.Errorf("token %s for %s does not exist", id, subject)
+		return nil, &store.ErrNotFound{Key: key}
 	}
 
 	claims := &types.Claims{}