+++ cli/client/testing/mock_authentication_client.go
@@ -8,6 +8,12 @@ func (c *MockClient) CreateAccessToken(url, u, p string) (*types.Tokens, error) 
 	return args.Get(0).(*types.Tokens), args.Error(1)
 }
 
+// TestCreds for use with mock lib
+func (c *MockClient) TestCreds(u, p string) error {
+	args := c.Called(u, p)
+	return args.Error(0)
+}
+
 // Logout for use with mock lib
 func (c *MockClient) Logout(token string) error {
 	args := c.Called(token)